+++ docs/source/changelog.rst
@@ -5,6 +5,8 @@ Change Log
 **Pyro 4.56**
 
 - serpent version check fix.
+- [on branch: dispatch_blob] experimental feature to be able to pass through serialized arguments unchanged via core.SerializedBlob
+  See example 'blob-dispatch'
 
 
 **Pyro 4.55**